requirements.txt
aiofiles==23.2.1
aioshutil==1.3
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
cffi==1.15.1
click==8.1.7
cloudpickle==2.2.1
cryptography==41.0.4
dask==2023.5.0
exceptiongroup==1.1.3
fastapi==0.103.1
fsspec==2023.9.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
locket==1.0.0
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.24.4
packaging==23.1
pandas==2.0.3
partd==1.4.0
passlib==1.7.4
pycparser==2.21
pydantic==1.10.9
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
zipp==3.17.0
